# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "VibeRoute API",
    "version": "0.1.0"
  },
  "paths": {
    "/v1/health": {
      "get": {
        "summary": "Health",
        "operationId": "health_v1_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/regions": {
      "get": {
        "summary": "Regions",
        "operationId": "regions_v1_regions_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/v1/provinces": {
      "get": {
        "summary": "Provinces",
        "operationId": "provinces_v1_provinces_get",
        "parameters": [
          {
            "name": "region",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Region"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/search": {
      "post": {
        "summary": "Search",
        "operationId": "search_v1_search_post",
        "parameters": [
          {
            "name": "X-Session-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/chats": {
      "get": {
        "summary": "List Chats",
        "operationId": "list_chats_v1_chats_get",
        "parameters": [
          {
            "name": "X-Session-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/chats/{chat_id}": {
      "get": {
        "summary": "Get Chat",
        "operationId": "get_chat_v1_chats__chat_id__get",
        "parameters": [
          {
            "name": "chat_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Chat Id"
            }
          },
          {
            "name": "X-Session-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/feedback": {
      "post": {
        "summary": "Feedback",
        "operationId": "feedback_v1_feedback_post",
        "parameters": [
          {
            "name": "X-Session-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/places/{att_id}/images": {
      "get": {
        "summary": "List Images",
        "operationId": "list_images_v1_places__att_id__images_get",
        "parameters": [
          {
            "name": "att_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Att Id"
            }
          },
          {
            "name": "X-Session-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Upload Image",
        "operationId": "upload_image_v1_places__att_id__images_post",
        "parameters": [
          {
            "name": "att_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Att Id"
            }
          },
          {
            "name": "X-Session-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_image_v1_places__att_id__images_post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaceImage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/images/{image_id}/favorite": {
      "post": {
        "summary": "Toggle Favorite",
        "operationId": "toggle_favorite_v1_images__image_id__favorite_post",
        "parameters": [
          {
            "name": "image_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Image Id"
            }
          },
          {
            "name": "X-Session-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaceImage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AssistantStatus": {
        "properties": {
          "level": {
            "type": "string",
            "enum": [
              "ready",
              "fallback",
              "off"
            ],
            "title": "Level"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "level",
          "label",
          "detail"
        ],
        "title": "AssistantStatus"
      },
      "Body_upload_image_v1_places__att_id__images_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_image_v1_places__att_id__images_post"
      },
      "FactField": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "confirmed",
              "unknown"
            ],
            "title": "Status"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "status",
          "label"
        ],
        "title": "FactField"
      },
      "FeedbackRequest": {
        "properties": {
          "message_id": {
            "type": "string",
            "title": "Message Id"
          },
          "att_id": {
            "type": "string",
            "title": "Att Id"
          },
          "rating": {
            "type": "integer",
            "enum": [
              1,
              -1
            ],
            "title": "Rating"
          }
        },
        "type": "object",
        "required": [
          "message_id",
          "att_id",
          "rating"
        ],
        "title": "FeedbackRequest"
      },
      "FeedbackResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "att_id": {
            "type": "string",
            "title": "Att Id"
          },
          "rating": {
            "type": "integer",
            "enum": [
              1,
              -1
            ],
            "title": "Rating"
          }
        },
        "type": "object",
        "required": [
          "ok",
          "att_id",
          "rating"
        ],
        "title": "FeedbackResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HealthResponse": {
        "properties": {
          "assistant": {
            "$ref": "#/components/schemas/AssistantStatus"
          },
          "embed_ready": {
            "type": "boolean",
            "title": "Embed Ready"
          },
          "listing_count": {
            "type": "integer",
            "title": "Listing Count"
          },
          "poc_region": {
            "type": "string",
            "title": "Poc Region"
          }
        },
        "type": "object",
        "required": [
          "assistant",
          "embed_ready",
          "listing_count",
          "poc_region"
        ],
        "title": "HealthResponse"
      },
      "MapPoint": {
        "properties": {
          "att_id": {
            "type": "string",
            "title": "Att Id"
          },
          "name_th": {
            "type": "string",
            "title": "Name Th"
          },
          "lat": {
            "type": "number",
            "title": "Lat"
          },
          "lng": {
            "type": "number",
            "title": "Lng"
          }
        },
        "type": "object",
        "required": [
          "att_id",
          "name_th",
          "lat",
          "lng"
        ],
        "title": "MapPoint"
      },
      "Place": {
        "properties": {
          "att_id": {
            "type": "string",
            "title": "Att Id"
          },
          "name_th": {
            "type": "string",
            "title": "Name Th"
          },
          "province": {
            "type": "string",
            "title": "Province"
          },
          "district": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "District"
          },
          "type_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Type Label"
          },
          "why": {
            "type": "string",
            "title": "Why"
          },
          "score_vector": {
            "type": "number",
            "title": "Score Vector"
          },
          "score_ranked": {
            "type": "number",
            "title": "Score Ranked"
          },
          "fee": {
            "$ref": "#/components/schemas/FactField"
          },
          "hours": {
            "$ref": "#/components/schemas/FactField"
          },
          "tel": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tel"
          },
          "website": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Website"
          },
          "facebook": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Facebook"
          },
          "limitation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Limitation"
          },
          "lat": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lat"
          },
          "lng": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lng"
          },
          "images": {
            "items": {
              "$ref": "#/components/schemas/PlaceImage"
            },
            "type": "array",
            "title": "Images"
          }
        },
        "type": "object",
        "required": [
          "att_id",
          "name_th",
          "province",
          "why",
          "score_vector",
          "score_ranked",
          "fee",
          "hours"
        ],
        "title": "Place"
      },
      "PlaceImage": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "att_id": {
            "type": "string",
            "title": "Att Id"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "fav_count": {
            "type": "integer",
            "title": "Fav Count",
            "default": 0
          },
          "is_cover": {
            "type": "boolean",
            "title": "Is Cover",
            "default": false
          },
          "viewer_faved": {
            "type": "boolean",
            "title": "Viewer Faved",
            "default": false
          },
          "moderation_status": {
            "type": "string",
            "enum": [
              "accepted",
              "pending",
              "rejected"
            ],
            "title": "Moderation Status",
            "default": "accepted"
          }
        },
        "type": "object",
        "required": [
          "id",
          "att_id",
          "url"
        ],
        "title": "PlaceImage"
      },
      "SearchRequest": {
        "properties": {
          "query": {
            "type": "string",
            "minLength": 1,
            "title": "Query"
          },
          "region": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region"
          },
          "province": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Province"
          },
          "prefer_secondary": {
            "type": "boolean",
            "title": "Prefer Secondary",
            "default": true
          },
          "chat_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chat Id"
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "SearchRequest"
      },
      "SearchResponse": {
        "properties": {
          "chat_id": {
            "type": "string",
            "title": "Chat Id"
          },
          "message_id": {
            "type": "string",
            "title": "Message Id"
          },
          "intro": {
            "type": "string",
            "title": "Intro"
          },
          "assistant": {
            "$ref": "#/components/schemas/AssistantStatus"
          },
          "prefer_secondary": {
            "type": "boolean",
            "title": "Prefer Secondary"
          },
          "secondary_count": {
            "type": "integer",
            "title": "Secondary Count"
          },
          "places": {
            "items": {
              "$ref": "#/components/schemas/Place"
            },
            "type": "array",
            "title": "Places"
          },
          "map_points": {
            "items": {
              "$ref": "#/components/schemas/MapPoint"
            },
            "type": "array",
            "title": "Map Points"
          }
        },
        "type": "object",
        "required": [
          "chat_id",
          "message_id",
          "intro",
          "assistant",
          "prefer_secondary",
          "secondary_count",
          "places",
          "map_points"
        ],
        "title": "SearchResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}